# backend/school-management/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management API",
        "description": "Documentación de la API para el sistema de gestión escolar",
        "contact": {
            "name": "Equipo de desarrollo CBTa71",
            "email": "soporte@cbta71.edu.mx"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Servidor local"
        },
        {
            "url": "https://nginx-production-728f.up.railway.app/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/api/v1/concepts": {
            "get": {
                "tags": [
                    "Payment Concepts"
                ],
                "summary": "Listar conceptos de pago",
                "description": "Obtiene una lista paginada de conceptos de pago, con filtros opcionales por estado y control de caché.",
                "operationId": "f9d62ceeda29d557b295b481705f8256",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtra por estado (ej. 'activo', 'todos', etc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "activo"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página a obtener",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "forceRefresh",
                        "in": "query",
                        "description": "Si es true, fuerza actualización de caché",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de conceptos de pago obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "concepts": {
                                                    "allOf": [
                                                        {
                                                            "$ref": "#/components/schemas/PaginatedResponse"
                                                        },
                                                        {
                                                            "properties": {
                                                                "items": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "$ref": "#/components/schemas/DomainPaymentConcept"
                                                                    }
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payment Concepts"
                ],
                "summary": "Crear un nuevo concepto de pago",
                "description": "Crea un nuevo concepto de pago y lo asocia con las entidades correspondientes (carreras, semestres, estudiantes).",
                "operationId": "743f0533fc17f5c5ab407a7991420bb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePaymentConceptDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Concepto de pago creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "concept": {
                                                    "$ref": "#/components/schemas/DomainPaymentConcept"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Concepto de pago creado con éxito."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de datos."
                                        },
                                        "errors": {
                                            "description": "Errores de validación por campo",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto en los datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Conflicto en los datos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recurso no encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/concepts/{id}": {
            "put": {
                "tags": [
                    "Payment Concepts"
                ],
                "summary": "Actualizar un concepto de pago",
                "description": "Actualiza los datos de un concepto de pago existente. Todos los campos son opcionales (usar 'sometimes'), excepto el id en la ruta.",
                "operationId": "5255c45ecbf80e25ad1524e85dbb6094",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del concepto de pago",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePaymentConceptDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Concepto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "concept": {
                                                    "$ref": "#/components/schemas/DomainPaymentConcept"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Concepto de pago actualizado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación de datos"
                    },
                    "409": {
                        "description": "Conflicto en los datos"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payment Concepts"
                ],
                "summary": "Eliminar concepto de pago (físicamente)",
                "operationId": "a1ce3785ce0caf528b2a88675bdad7d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concepto de pago eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginatedResponse": {
                "properties": {
                    "items": {
                        "description": "Lista de elementos paginados",
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "currentPage": {
                        "type": "integer",
                        "example": 1
                    },
                    "lastPage": {
                        "type": "integer",
                        "example": 5
                    },
                    "perPage": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 72
                    },
                    "hasMorePages": {
                        "type": "boolean",
                        "example": true
                    },
                    "nextPage": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}